custom error class to middleware
