+++ helper.js
@@ -6,7 +6,17 @@ function getAccessCode(){
     return Math.floor(Math.random() * (max - min + 1)) + min;
 }
 
+function ValidateEmail(mail) 
+{
+ if (/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(mail))
+  {
+    return (true)
+  }
+    return (false)
+}
+
 module.exports = 
 {
-    getAccessCode
+    getAccessCode,
+    ValidateEmail
 }
